+++ docs/source/news.rst
@@ -7,6 +7,7 @@ What's New
 Features:
 
 * ``eliot-prettyprint`` is a command-line tool that formats JSON Eliot messages into a more human-friendly format.
+* ``eliot.logwriter.ThreadedWriter`` is a Twisted non-blocking wrapper for any blocking destination.
 
 0.8.0
 ^^^^^